GetOrSetItem with lazy value factory
GetOrSetItem: If the item exists for a cache key, then return it from the cache without executing the value factory. If the item for that cache key doesn’t exist, then add the result of the value factory to the cache, then return it.

Inspired by the GetOrCreate method in Microsoft.Extensions.Caching.MemoryCache.Extensions.